Update radio button class.
