+++ Source/RadioButton/RadioButtonColor.swift
@@ -1,0 +1,23 @@
+//
+//  RadioButtonColor.swift
+//  RadioAndCheckboxButtonDemo
+//
+//  Created by Manish Bhande on 08/12/18.
+//  Copyright © 2018 Manish. All rights reserved.
+//
+
+import UIKit
+
+/// Define RadioButtonColor
+
+public struct RadioButtonColor {
+    
+    let active: UIColor
+    let inactive: UIColor
+    
+    public init(active: UIColor, inactive: UIColor) {
+        self.active = active
+        self.inactive = inactive
+    }
+    
+}
